Fix in qt clustering
Major fixes in weighted QT clustering.

Before the fix, clusters were searched according to the sum of
all weights except for the first member. This was leading to incorrect
results if there are structures with large weights.

An additional problem was that presence of structures with distance==0.0
was leading to errors. This is fixed now.

Results in the regtests are unaffected by this change (except
for the appearance of size=1 clusters, see below), but a
new regtest (tests/test_clustering.py test_qt5)
probes exactly these situations.

As a side effect, clusters of size 1 are now correctly found.
This affects also clustering without weights.